+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] 2020-12-08
+### Added
+- The nameplate 9.0.1 widget containers should now be visible on our nameplates too. No styling as of yet, we simply made them visible.
+
 ### Changed
 - Changed chat bubble font for latin clients and increased its size for better readability.
 - Added some updates to the class power element to try to battle the issue of weirdly looking points after cinematics.